Implement project management features: add project creator assignment, user-specific project access checks, and enhance project schema with manager field.

+++ src/controllers/ProjectController.ts
@@ -5,10 +5,8 @@ export class ProjectController {
   static createProject = async (req: Request, res: Response) => {
     const project = new Project(req.body);
 
-    /* if (true) {
-      const error = new Error("Proyecto no encontrado");
-      return res.status(404).json({ error: error.message });
-    } */
+    // Assign the creator of the project
+    project.manager = req.user?.id;
 
     try {
       await project.save();
@@ -20,7 +18,13 @@ export class ProjectController {
 
   static getAllProjects = async (req: Request, res: Response) => {
     try {
-      const projects = await Project.find({});
+      // Fetch all projects from the database by user ID
+      const projects = await Project.find({
+        $or: [{ manager: req.user?.id }]
+      });
+
+      //const projects = await Project.find({ manager: req.user?.id });
+
       return res.json(projects);
     } catch (error) {
       console.log(error);
@@ -36,6 +40,12 @@ export class ProjectController {
       if (!project) {
         const error = new Error("Proyecto no encontrado");
         return res.status(404).json({ error: error.message });
+      }
+
+      // Check if the project belongs to the user
+      if (project.manager.toString() !== req.user?.id.toString()) {
+        const error = new Error("Accion no valida");
+        return res.status(403).json({ error: error.message });
       }
 
       res.json(project);
@@ -55,6 +65,12 @@ export class ProjectController {
         return res.status(404).json({ error: error.message });
       }
 
+      // Check if the project belongs to the user
+      if (project.manager.toString() !== req.user?.id.toString()) {
+        const error = new Error("Solo el creador puede editar el proyecto");
+        return res.status(403).json({ error: error.message });
+      }
+
       project.clientName = req.body.clientName;
       project.projectName = req.body.projectName;
       project.description = req.body.description;
@@ -71,6 +87,18 @@ export class ProjectController {
 
     try {
       const project = await Project.findById(id);
+
+      if (!project) {
+        const error = new Error("Proyecto no encontrado");
+        return res.status(404).json({ error: error.message });
+      }
+
+      // Check if the project belongs to the user
+      if (project.manager.toString() !== req.user?.id.toString()) {
+        const error = new Error("Solo el creador puede eliminar el proyecto");
+        return res.status(403).json({ error: error.message });
+      }
+
       await project.deleteOne();
       res.send("Proyecto eliminado");
     } catch (error) {
